refactor: move gateway api structs into lnv2 common
