+++ modules/fedimint-lnv2-client/Cargo.toml
@@ -34,7 +34,6 @@ futures = { workspace = true }
 itertools = { workspace = true }
 lightning-invoice = { workspace = true }
 rand = { workspace = true }
-reqwest = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 thiserror = { workspace = true }
